Edited entity abstraction
