--- app/src/main/java/com/mtlepberghenov/internship_playground/model/entity/CarAction.java
@@ -1,8 +1,0 @@
-package com.mtlepberghenov.internship_playground.model.entity;
-
-public interface CarAction {
-
-    void switchOnEngine();
-
-    void switchOffEngine();
-}
