--- app/src/main/java/com/mtlepberghenov/internship_playground/model/entity/Toyota.java
@@ -1,18 +1,0 @@
-package com.mtlepberghenov.internship_playground.model.entity;
-
-public class Toyota extends Car implements CarAction {
-
-    public Toyota(String carMake, String model, String color, int year, CarKitting carKitting) {
-        super(carMake, model, color, year, carKitting);
-    }
-
-    @Override
-    public void switchOnEngine() {
-
-    }
-
-    @Override
-    public void switchOffEngine() {
-
-    }
-}
